move code around, update to latest x/sys/unix
