+++ formats.go
@@ -29,7 +29,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/tobert/pcstat"
+	"github.com/tobert/pcstat/pkg"
 )
 
 type PcStatusList []pcstat.PcStatus
