+++ go.mod
@@ -2,4 +2,4 @@ module github.com/tobert/pcstat
 
 go 1.13
 
-require golang.org/x/sys v0.0.0-20191127021746-63cb32ae39b2
+require golang.org/x/sys v0.0.0-20210917161153-d61c044b1678
